apps/notifications/apps.py
from __future__ import annotations
from django.apps import AppConfig

class NotificationsConfig(AppConfig):
    name = "apps.notifications"
    label = "notifications"
    verbose_name = "Notifications"
    default_auto_field = "django.db.models.BigAutoField"

    def ready(self) -> None:
        from apps.users.application.permissions import register_permissions
        register_permissions("notifications", ("view", "mark_read"))
